Minor optimizations
Store beta_s only in the ion beam for now to clear up the
ambiguity of storing it both in the ion beam and in the ring.
Having that variable as a property of the beam feels structurally
awkward, but some amount of awkwardness is currently unavoidable
because of the interdependence of Ring and IonBeam that cannot
be fixed easily or at all.

Also call IonBeam::create_samples() by default.